+++ tests/conf.js
@@ -65,7 +65,7 @@ const baseConfig = {
     browser.waitForAngularEnabled(false);
 
     // wait for startup to complete
-    await browser.driver.wait(prepServices(), 135 * 1000, 'API took too long to start up');
+    await browser.driver.wait(utils.prepServices(), 135 * 1000, 'API took too long to start up');
 
     afterEach(() => {
       return browser
@@ -93,48 +93,6 @@ const runAndLog = (msg, func) => {
   return func();
 };
 
-const prepServices = async () => {
-  if (constants.IS_TRAVIS) {
-    console.log('On travis, waiting for horti to first boot api');
-    // Travis' horti will be installing and then deploying api and sentinel, and those logs are
-    // getting pushed into horti.log Once horti has bootstrapped we want to restart everything so
-    // that the service processes get restarted with their logs separated and pointing to the
-    // correct logs for testing
-    await listenForApi();
-    console.log('Horti booted API, rebooting under our logging structure');
-    await request.post('http://localhost:31337/all/restart');
-  } else {
-    // Locally we just need to start them and can do so straight away
-    await request.post('http://localhost:31337/all/start');
-  }
-
-  await listenForApi();
-  const config = await browser.getProcessedConfig();
-  if (config.suite && config.suite === 'web'){
-    await runAndLog('Settings setup', setupSettings);
-  }
-  await runAndLog('User contact doc setup', utils.setUserContactDoc);
-};
-
-const apiRetry = () => {
-  return new Promise(resolve => {
-    setTimeout(() => {
-      resolve(listenForApi());
-    }, 1000);
-  });
-};
-
-const listenForApi = async () => {
-  console.log('Checking API');
-  try {
-    await utils.request({ path: '/api/info' });
-  } catch(err) {
-    console.log('API check failed, trying again in 1 second');
-    console.log(err.message);
-    await apiRetry();
-  }
-};
-
 const getLoginUrl = () => {
   const redirectUrl = encodeURIComponent(
     `/${constants.DB_NAME}/_design/${constants.MAIN_DDOC_NAME}/_rewrite/#/messages`
@@ -155,17 +113,6 @@ const login = async browser => {
     20 * 1000,
     'Login should be complete within 20 seconds'
   );
-};
-
-const setupSettings = () => {
-  const defaultAppSettings = utils.getDefaultSettings();
-  defaultAppSettings.transitions = {};
-
-  return utils.request({
-    path: '/api/v1/settings?replace=1',
-    method: 'PUT',
-    body: defaultAppSettings
-  });
 };
 
 const setupUser = () => {
